LayoutTests/js/Promise-simple-fulfill.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise.');

window.jsTestIsAsync = true;

var future = new Promise(function(resolver) {
  setTimeout(function() {
    resolver.fulfill('hello');
  }, 0);
}).then(function(result) {
  window.result = result;
  shouldBeEqualToString('result', 'hello');
  finishJSTest();
});

</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
